Refactor user_profile service
